feat(LiquidationRewards): Change liquidation rewards multiplier denominator to a 0.01% precision

+++ src/interfaces/OracleMiddleware/ILiquidationRewardsManager.sol
@@ -31,6 +31,6 @@ interface ILiquidationRewardsManager is ILiquidationRewardsManagerErrorsEventsTy
         uint32 gasUsedPerTick,
         uint32 otherGasUsed,
         uint64 gasPriceLimit,
-        uint16 multiplierBps
+        uint32 multiplierBps
     ) external;
 }
